Change teachers id to int + removed repo tests

+++ src/main/java/com/springbootproject/controller/TeacherController.java
@@ -40,13 +40,13 @@ public class TeacherController {
     }
 
     @GetMapping("/teachers/edit/{id}")
-    public String editTeacherForm(@PathVariable Long id, Model model) {
+    public String editTeacherForm(@PathVariable int id, Model model) {
         model.addAttribute("teacher", teacherService.getTeacherById(id));
         return "teacher/edit_teacher";
     }
 
     @PostMapping("/teachers/{id}")
-    public String updateTeacher(@PathVariable Long id,
+    public String updateTeacher(@PathVariable int id,
                                 @ModelAttribute("teacher") Teacher teacher,
                                 Model model) {
         //get teacher from database by id
@@ -61,7 +61,7 @@ public class TeacherController {
     }
 
     @GetMapping("teachers/{id}")
-    public String deleteTeacher(@PathVariable Long id) {
+    public String deleteTeacher(@PathVariable int id) {
         teacherService.deleteTeacherById(id);
         return "redirect:/teachers";
     }
